// src/app/models/IInformeRechazoMensual.ts
export interface IInformeRechazoMensual {
    componente: string;
    subComponente: string;
    defecto: string;
    totalDefectos: number;
    fecha: string;
    hora: string;
    linea: string;
    barcode: string;
    puesto: string;
    estadoRechazo: string;
    lineaId: number;
    idCodigoRechazo: number;
    turno: string;
    familia: string;
    rechazoDescripcion: string;
    totaliza: string
}